// src/components/ads-check/ads-check.interface.ts
export interface AdsCheck {
  title: string | null;
  url: string | null;
  square?: string | null;
  floor?: string | null;
  maxFloor?: string | null;
  smallDescription?: string | null;
  sellerName?: string | null;
  street?: string | null;
  price?: string | null;
  house?: string | null;
  datePublished: string | null;
  hash: string;
}